refactor: reorganize middleware imports and remove unused decorators
