--- telegram_bot/src/aiogram/decorators/decorators.py
@@ -1,8 +1,0 @@
-from functools import wraps
-
-from aiogram.types import Message
-from aiogram import Bot
-
-from src.database import Database
-
-# Empty
