+++ telegram_bot/src/aiogram/handlers/commands.py
@@ -6,7 +6,7 @@ from aiogram.enums import ParseMode
 from src.logger import logger
 
 from src.database import Redis, Database
-from src.aiogram.middlewares.middlewares import WaitingMiddleware, CheckNewUserMiddleware
+from src.aiogram.middlewares import WaitingMiddleware, CheckNewUserMiddleware
 from src.config import TRIAL_PERIOD_NUM_REQ
 from src.aiogram.utils import commands_text, answer_message
 from src.gpt import OpenAI_API
@@ -42,40 +42,34 @@ async def reset_handler(message: Message):
     await message.answer(text, parse_mode=ParseMode.MARKDOWN_V2)
 
 
-
 @router.message(Command('show_dialog'), ChatModeFilter(mode="usual"))
 async def show_dialog_handler(message: Message, redis: Redis):
     history = await redis.get_history(message.from_user.id)
     if not history:
         await message.answer("Диалог пуст")
         return
+
     for cur_message in history:
-        sender = "Неизветно кто"
+        sender = "Неизвестно кто"
         if cur_message["role"] == "user":
             sender = "Пользователь"
-
             mes = ""
 
-            content = cur_message['content']
+            for cur_content in cur_message['content']:
+                if cur_content.get("type") == "text":
+                    mes += f"{cur_content['text']}\n\n"
+                elif cur_content.get("type") == "image_url":
+                    mes += "_Приложено фото_"
 
-            for cur_content in content:
-                if "type" in cur_content:
-                    if cur_content["type"] == "text":
-                        mes += f"{cur_content["text"]}\n\n"
-                    if cur_content["type"] == "image_url": 
-                        mes += "_Приложено фото_"
-            
             await answer_message(
-                md=f"*{sender}*:\n" + mes,
+                md=f"*{sender}*:\n{mes}",
                 message=message,
             )
 
-
         elif cur_message["role"] == "assistant":
             sender = "Бот"
-
             await answer_message(
-                md=f"*{sender}*:\n" + cur_message['content'],
+                md=f"*{sender}*:\n{cur_message['content']}",
                 message=message,
             )
