+++ telegram_bot/src/aiogram/handlers/unknown.py
@@ -1,10 +1,7 @@
-from aiogram import Router, F
-from aiogram.types import Message, ContentType
+from aiogram import Router
+from aiogram.types import Message
 
-from src.aiogram.middlewares.middlewares import WaitingMiddleware, CheckNewUserMiddleware
-
-from src.logger import logger
-
+from src.aiogram.middlewares import WaitingMiddleware, CheckNewUserMiddleware
 
 router = Router()
