+++ telegram_bot/src/aiogram/middlewares/middlewares.py
@@ -25,118 +25,7 @@ from src.aiogram.utils import commands_text, is_sha256, vk_send_pixel_event
 
 from src.logger import logger
 
-
-class ErrorLoggingMiddleware(BaseMiddleware):
-    def __init__(self, bot: Bot, db: Database, redis: Redis):
-        self.bot = bot
-        self.db = db
-        self.redis = redis
-
-    async def __call__(self, handler, event: Update, data: dict):
-        try:
-            # Выполняем основной обработчик
-            return await handler(event, data)
-        except Exception as exception:
-            # Получаем информацию о пользователе
-            telegram_id = None
-            if event.message:
-                telegram_id = event.message.from_user.id
-            elif event.callback_query:
-                telegram_id = event.callback_query.from_user.id
-
-            # Получаем информацию об ошибке
-            tb = traceback.extract_tb(exception.__traceback__)
-            filepath, lineno, func_name, line = tb[-1]
-
-            # Логируем ошибку
-            logger.error(
-                f"Global error occurred\n{filepath}\nFunc name: {func_name}\n{exception.__class__.__name__} | {exception}",
-                exc_info=False
-            )
-
-            # Логируем ошибку в базу данных
-            await self.db.add_error(
-                error_type=str(exception.__class__.__name__),
-                error_text=str(exception),
-                file_path=filepath,
-                telegram_id=telegram_id,
-                traceback=traceback.format_exc()
-            )
-
-            # Удаляем историю сессии из Redis
-            if telegram_id:
-                await self.redis.clear_user_history(telegram_id)
-                await self.redis.set_user_req_inactive(telegram_id)
-
-                # Уведомляем пользователя
-                await self.bot.send_message(
-                    chat_id=telegram_id,
-                    text=f"Произошла непредвиденная ошибка\nСвяжитесь с разработчиком (@yurchest)\nError: {exception}"
-                )
-
-            # Блокируем выполнение
-            # raise exception
-
-class TimingMessageMiddleware(BaseMiddleware):
-    async def __call__(self, handler, event: TelegramObject, data: dict):
-        """
-        Замеряем MESSAGE_RESPONSE_TIME
-        """
-        start_time = datetime.now()  # Фиксируем момент получения сообщения ботом
-
-        result =  await handler(event, data)
-        
-        if isinstance(event, Message):
-            # Время обработки сообщения основной логикой
-            response_time = (datetime.now() - start_time).total_seconds() 
-            # Время от отправки сообщения пользователем до начала основной логики
-            message_latency: datetime = (start_time - event.date.replace(tzinfo=None)).total_seconds()  
-            # Общее время от отправки до ответа 
-            total_latency = message_latency + response_time  
-            # Логируем
-            logger.debug(f"(MAIN)\t\t Telegram latency: {message_latency:.3f}s, Bot processing: {response_time:.3f}s, Total: {total_latency:.3f}s")
-            # Пишем в Prometheus
-            MESSAGE_RESPONSE_TIME.observe(total_latency) # Отправляем в Prometheus
-        
-        return result
-
-        
-
-class DatabaseMiddleware(BaseMiddleware):
-    def __init__(self, db: Database):
-        super().__init__()
-        self.db = db
-
-    async def __call__(self, handler, event: TelegramObject, data: dict):
-        """
-        Добавляет объект `db` в `data`, чтобы он был доступен в хендлерах.
-        """
-        data["db"] = self.db
-        return await handler(event, data)
     
-class OpenAIMiddleware(BaseMiddleware):
-    def __init__(self, openai: OpenAI_API):
-        super().__init__()
-        self.openai = openai
-
-    async def __call__(self, handler, event: TelegramObject, data: dict):
-        """
-        Добавляет объект `openai` в `data`, чтобы он был доступен в хендлерах.
-        """
-        data["openai"] = self.openai
-        return await handler(event, data)
-    
-class RedisMiddleware(BaseMiddleware):
-    def __init__(self, redis: Redis):
-        super().__init__()
-        self.redis = redis
-
-    async def __call__(self, handler, event: TelegramObject, data: dict):
-        """
-        Добавляет объект `redis` в `data`, чтобы он был доступен в хендлерах.
-        """
-        data["redis"] = self.redis
-        return await handler(event, data)
     
     
 class CheckNewUserMiddleware(BaseMiddleware):
@@ -205,10 +94,16 @@ class IncrementRequestsMiddleware(BaseMiddleware):
         if isinstance(event, Message):
             # Получаем объект базы данных из контекста
             db = data.get("db")
+            redis: Redis = data.get("redis")
             
             if db is None:
                 raise ValueError("Database instance must be provided in the context data.")
+            if redis is None:
+                raise ValueError("Redis instance must be provided in the context data.")
             
+            # Если первый запрос,отправляем событие в ВК рекламу
+            if await db.get_num_requests(event.from_user.id) == 0:
+                await vk_send_pixel_event(redis=redis, user_id=event.from_user.id, goal_name="first_requset", cost=20)
             # Увеличиваем счетчик запросов пользователя
             await db.increment_user_requests(event.from_user.id)
             # Обновляем дату последнего запроса
@@ -287,49 +182,6 @@ class CheckSubscriptionMiddleware(BaseMiddleware):
             return await handler(event, data)
 
     
-class WaitingMiddleware(BaseMiddleware):
-    """
-    Если предыдущий запрос пользователя еще обрабатывается, отправляет техническое сообщение об этом.
-    Иначе выводим техническое сообщение - точки.
-    В процессе запроса присваиваем активность (Redis).
-    После выполнения технические сообщения удаляются.
-    """ 
-
-    async def __call__(self, handler, event: TelegramObject, data: dict):
-        # Получаем объект базы данных из контекста
-        redis: Redis = data.get("redis")
-
-        if redis is None:
-            raise ValueError("Redis instance must be provided in the context data.")
-        
-
-        # Проверяем, активен ли запрос пользователя
-        is_user_waiting = await redis.is_user_waiting(event.from_user.id)
-        
-
-        tech_message = None  # Для хранения ссылки на отправленное сообщение
-
-        if is_user_waiting:
-            # Если запрос пользователя активен, отправляем сообщение о том, что запрос обрабатывается
-            tech_message = await event.answer("Ваш запрос обрабатывается. Пожалуйста, подождите...")
-            asyncio.create_task(delete_message_when_inactive(redis, event.from_user.id, tech_message, event))
-            return  # Завершаем выполнение, так как запрос уже обрабатывается
-        
-        
-        # Отправляем техническое сообщение с точками
-        tech_message = await event.answer(". . . . . .")
-        # Устанавливаем флаг активности запроса пользователя
-        await redis.set_user_req_active(event.from_user.id)
-        # Ожидаем когда запрос станет неактивным
-        asyncio.create_task(delete_message_when_inactive(redis, event.from_user.id, tech_message))
-        # Вызываем следующий обработчик
-        result = await handler(event, data)
-        # Удаляем флаг активности запроса пользователя
-        await redis.set_user_req_inactive(event.from_user.id)
-
-        
-        return result
-
 
 
 class CheckHistoryLengthMiddleware(BaseMiddleware):
@@ -416,25 +268,6 @@ class CheckHistoryLengthMiddleware(BaseMiddleware):
 
         return result
 
-
-async def delete_message_when_inactive(
-            redis: Redis, 
-            user_id: int,
-            tech_message: TelegramObject, 
-            user_message: TelegramObject = None
-            ):
-        """
-        Ожидает, пока пользователь ждет ответа.
-        """
-        # Удаляем пользователское сообщение
-        if user_message: await user_message.delete()
-
-        while await redis.is_user_waiting(user_id):
-            # Ожидаем пока юзеру ответит бот на предыдущее сообщение
-            await asyncio.sleep(0.1)
-
-        # Удаляем техническое сообщение
-        await tech_message.delete()
